+++ CHANGELOG.md
@@ -10,6 +10,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 To see unreleased changes, please see the [CHANGELOG on the master branch](https://github.com/gufolabs/gufo_err/blob/master/CHANGELOG.md) guide.
 
+## [Unreleased]
+
+### Fixed
+
+* Suppress Sentry SDK traceback on tests exit.
+
 ## 0.6.0 - 2025-10-10
 
 ### Added